move useIsVisible hook to separate file
