+++ src/Overview/OverviewHeader.tsx
@@ -3,7 +3,7 @@ import { jupyterIcon, LabIcon } from '@jupyterlab/ui-components';
 import React, { useEffect, useRef, useState } from 'react';
 import { LoopsLogo } from '../assets/loops-logo';
 import { createStyles, Dialog, Modal, Text } from '@mantine/core';
-import { useIsVisible } from '../util';
+import { useIsVisible } from '../useIsVisible';
 import { useDisclosure } from '@mantine/hooks';
 
 const useStyles = createStyles((theme, _params) => ({
