+++ src/util.ts
@@ -1,5 +1,3 @@
-import { useEffect, useState } from 'react';
-
 /**
  * Merge arrays, preserving order of elements
  * @param newArray defines the order of elements
@@ -118,18 +116,3 @@ export const getScrollParent = (node: Element): Element => {
   }
   return document.scrollingElement || document.documentElement;
 };
-
-export function useIsVisible(ref) {
-  const [isIntersecting, setIntersecting] = useState(false);
-
-  useEffect(() => {
-    const observer = new IntersectionObserver(([entry]) => setIntersecting(entry.isIntersecting));
-
-    observer.observe(ref.current);
-    return () => {
-      observer.disconnect();
-    };
-  }, [ref]);
-
-  return isIntersecting;
-}
